change app constant color directory
