--- lib/constants/colors.dart
@@ -1,50 +1,0 @@
-import 'package:flutter/widgets.dart';
-
-abstract class AppConstantColor {
-  // light theme constant color
-  static const Color lightThemeBackgroundColor = Color(0xffF2F2F2);
-  static const Color lightThemeBottomNavColor = Color(0xffFFFFFF);
-  static const Color lightThemeBottomNavShadowColor = Color(0x4A3C3C43);
-  static const Color lightThemePrimaryTextColor = Color(0xff262626);
-  static const Color lightThemeSecondaryTextColor = Color(0x66000000);
-  static const Color lightThemeTextFiledColor = Color(0x1A000000);
-  static const Color lightThemeTextFiledLabelColor = Color(0x33000000);
-  static const Color lightThemeIconColor = Color(0xff000000);
-  static const Color lightThemeAppBarColor = Color(0xffFAFAFA);
-  static const Color lightThemeAppBarShowColor = Color(0xffA6A6AA);
-  static const Color lightThemePostInActiveIndicatorColor = Color(0x26000000);
-  static const Color lightThemeSearchBarBackgroundColor = Color(0x1F767680);
-  static const Color lightThemeSearchBarForegroundColor = Color(0x993C3C43);
-  static const Color lightThemeExploreCategoryTextColor = Color(0xff3C3C43);
-  static const Color lightThemeExploreCategoryBorderColor = Color(0x2E3C3C43);
-  static const Color lightThemeDividerColor = Color(0x4A3C3C43);
-  static const Color lightThemeViewedStoryBorder = Color(0xff262626);
-  // dark theme constant color
-  static const Color darkThemeBackgroundColor = Color(0xff000000);
-  static const Color darkThemeBottomNavColor = Color(0xff000000);
-  static const Color darkThemeBottomNavShadowColor = Color(0x26FFFFFF);
-  static const Color darkThemePrimaryTextColor = Color(0xffffffff);
-  static const Color darkThemeSecondaryTextColor = Color(0x99ffffff);
-  static const Color darkThemeTextFiledColor = Color(0xff121212);
-  static const Color darkThemeTextFiledLabelColor = Color(0x99FFFFFF);
-  static const Color darkThemeIconColor = Color(0xffF9F9F9);
-  static const Color darkThemeAppBarColor = Color(0xff121212);
-  static const Color darkThemeAppBarShowColor = Color(0x26FFFFFF);
-  static const Color darkThemePostInActiveIndicatorColor = Color(0x54ffffff);
-  static const Color darkThemeSearchBarBackgroundColor = Color(0xff262626);
-  static const Color darkThemeSearchBarForegroundColor = Color(0xff8E8E93);
-  static const Color darkThemeExploreCategoryTextColor = Color(0xffF9F9F9);
-  static const Color darkThemeExploreCategoryBorderColor = Color(0x26ffffff);
-  static const Color darkThemeDividerColor = Color(0x38000000);
-  static const Color darkThemeProfileBackgroundColor = Color(0xff121212);
-  static const Color darkThemeViewedStoryBorder = Color(0x1A000000);
-  // common constant color
-  static const Color postIndexWidgetColor = Color(0xB3121212);
-  static const Color postIndexTextWidgetColor = Color(0xffF9F9F9);
-  static const Color postActiveIndicatorColor = Color(0xffA6A6AA);
-  static const Color linkTextColor = Color(0xff05386B);
-  static const Color postForegroundIconColor = Color(0xffFEFFFE);
-  static const Color buttonBackgroundColor = Color(0xff3797EF);
-  static const Color buttonForegroundColor = Color(0xffFFFFFF);
-  static const Color textButtonColor = Color(0xff3797EF);
-}
